made limits for provenance graphs a little larger

+++ common/src/java/gov/fnal/elab/util/ElabUtil.java
@@ -474,8 +474,8 @@ public class ElabUtil {
             // Thanks to the Batik website's tutorial for this code
             // (http://xml.apache.org/batik/rasterizerTutorial.html).
             PNGTranscoder t = new PNGTranscoder();
-            t.addTranscodingHint(PNGTranscoder.KEY_MAX_HEIGHT, new Float(800));
-            t.addTranscodingHint(PNGTranscoder.KEY_MAX_WIDTH, new Float(1200));
+            t.addTranscodingHint(PNGTranscoder.KEY_MAX_HEIGHT, new Float(1200));
+            t.addTranscodingHint(PNGTranscoder.KEY_MAX_WIDTH, new Float(1400));
             TranscoderInput input = new TranscoderInput((new File(svg)).toURL()
                     .toString());
             OutputStream ostream = new FileOutputStream(png);
